requirements.txt
streamlit
sentence-transformers
faiss-cpu
chromadb
openai
whisper
vosk
moviepy
youtube_dl
langchain
python-dotenv
